% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\name{as.spectra.database}
\alias{as.spectra.database}
\title{Create a Spectra or SpectraDatabase}
\usage{
as.spectra.database(spectra = matrix(0), wavelength = numeric(0),
  w.unit = "nm", data = data.frame(), ...)
}
\description{
This function creates a Spectra object.
}
