% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.nsr.R
\name{cm.nsr}
\alias{cm.nsr}
\alias{plot.cm}
\title{Selecting the best 2-Band combinations for Normalized Simple Ratio (NSR)}
\usage{
cm.nsr(spectra, x, w = wavelength(spectra), w.unit = NULL,
  cm.plot = FALSE)

plot.cm(cm)
}
\arguments{
\item{spectra}{A matrix of spectral data, a row is a spectrum across all spectral bands.}

\item{x}{A vector.}

\item{w}{A vector of wavelength.}

\item{cm.plot}{A logic value for whether plotting the coefficient matrix or not, default FALSE.}

\item{cm}{A square matrix}
}
\value{
\item{cm}{Returns a coorrelation coefficients matrix corresponding to each band combination for NDVI.}

\item{cmp}{Returns a coorrelation-matrix plot.}
}
\description{
This function developes a optimization algorithm based on correlation analysis between spectral matrix 'spectra' and the
getation variable of interest x, which
determines the best spectral band combinations of the full spectrum that are most predictive for 'x'.
}
\details{
This function runs a calculation of \deqn{ NDVI = (\lambda_i - \lambda_j)/(\lambda_i + \lambda_j) } using all the possible pairs/combinations of any two bands (i,j)
within the full spectrum range thoroughly. A correlation analysis is then performed between the x and all possible NDVIs, and it calculates
the correlation coefficients (r) which indicates the predictive performance of each NDVI and its corresponding two-band combination. The
output is the wavelength (nm) indicating the best two bands that produce the highest value of r.
}
\examples{
  \dontrun{
  data(NSpec.DF)
  x <- NSpec.DF$N # nitrogen
  S <- NSpec.DF$spectra[, seq(1, ncol(NSpec.DF$spectra), 5)] # resampled to 10 nm steps
  cm.nsr(S, x, cm.plot = TRUE)
  }
}
\seealso{
\code{\link{cor}}
}
