% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vi.R
\name{compute_vi}
\alias{compute_vi}
\title{Calculate Vegetation Indices}
\usage{
compute_vi(x)
}
\arguments{
\item{x}{matrix or dataframe with columns for wavebands and rows for observations,
colnames should contain wavelength, rownames can be used for sample IDs.}
}
\value{
\item{v}{returns a dataframe of vegetation indices}
}
\description{
This function calculated vegetation indices used in remote sensing studies
and applications.
}
\references{

}
