% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.bestBands.R
\name{find.bestBands}
\alias{find.bestBands}
\title{Find Best Band Combinations}
\usage{
find.bestBands(R, w)
}
\arguments{
\item{R}{A numeric matrix (2D) or array (3D) containing metric values (e.g., correlation values).}

\item{w}{A numeric vector of wavelengths corresponding to the bands.}
}
\value{
A vector of wavelengths corresponding to the best band combination.
}
\description{
This function identifies the best band combination from the \link{cm.nsr} or \link{cm.sr}
calculated correlation coefficients - a numeric matrix (2D), or the \link{cm.rbd3} returned array (3D)
by locating the maximum value in the input \code{R} and returning the corresponding wavelengths
from \code{w}. For a 2D matrix, it returns a two-element vector (i, j); for a 3D array, a three-element
vector (i, j, k).
}
\details{
The function first verifies that \code{R} has dimensions. It then computes the maximum value in \code{R},
retrieves the indices corresponding to that value, and extracts the wavelengths from \code{w} based on the
dimensionality of \code{R}. If \code{R} is 2D, the order is assumed to be (i, j); if \code{R} is 3D, the order
is (i, j, k).
}
\examples{
# Example for a 2D matrix:
R_mat <- matrix(c(0.2, 0.8, 0.5, 0.3), nrow = 2)
wavelengths <- c(450, 550)
bestBands <- find.bestBands(R_mat, wavelengths)

# Example for a 3D array:
R_arr <- array(runif(27), dim = c(3, 3, 3))
wavelengths <- c(400, 450, 500)
bestBands <- find.bestBands(R_arr, wavelengths)

}
