% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.pair.R
\name{ndvi.pair}
\alias{ndvi.pair}
\title{Plotting the best NDVI calculated using \code{\link{ndvi}}}
\usage{
ndvi.pair(x, y, w, p)
}
\arguments{
\item{x}{A matrix of spectral data, a row is a spectrum across all spectral bands.}

\item{y}{A vector.}

\item{w}{A vector of wavelength.}

\item{p}{A vector of two elements (the pair of bands used for NDVI calculation).}
}
\description{
Plotting the best NDVI calculated using \code{\link{ndvi}}
}
\details{
This function plots the best NDVI against the y
}
\examples{
library(visa)
y <- specDF$N
x <- specDF$Spec
w <- as.numeric(gsub("nm", "", colnames(specDF$Spec)))
p <- c(440,445)
ndvi.pair(x,y,w,p)
}
