% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Selecting the best 2-Band combinations for Normalized Difference Vegetation Index (NDVI)}
\usage{
ndvi(X, y, w, plot.cm = FALSE)
}
\arguments{
\item{X}{A matrix of spectral data, a row is a spectrum across all spectral bands.}

\item{y}{A vector.}

\item{w}{A vector of wavelength.}

\item{plot.cm}{A logic value for whether plotting the coefficient matrix or not, default FALSE.}
}
\value{
\item{cm}{Returns a coorrelation coefficients matrix corresponding to each band combination for NDVI.}
}
\description{
This function developes a optimization algorithm based on correlation analysis between spectral parameter 'X' and y, which
determines the best spectral bands of the full spectrum that are most predictive for 'y'.
}
\details{
This function runs a calculation of \deqn{ NDVI = (\lambda_i - \lambda_j)/(\lambda_i + \lambda_j) } using all the possible pairs/combinations of any two bands (i,j)
within the full spectrum range thoroughly. A correlation analysis is then performed between the y and all possible NDVIs, and it calculates
the correlation coefficients (r) which indicates the predictive performance of each NDVI and its corresponding two-band combination. The
output is the wavelength (nm) indicating the best two bands that produce the highest value of r.
}
\examples{
library(visa)
y <- specDF$N # nitrogen
X <- specDF$Spec[, seq(1, ncol(specDF$Spec), 5)] # resampled to 5 nm steps
w <- gsub("nm", "", colnames(specDF$Spec))
w <- as.numeric(w)[seq(1, ncol(specDF$Spec), 5)] # resampled to 5 nm steps
ndvi(X,y,w)
}
\seealso{
\code{\link{cor}}
}
