% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi2.R
\name{ndvi2}
\alias{ndvi2}
\title{Calculate and plot a 2-band NDVI.}
\usage{
ndvi2(s, b1, b2)
}
\arguments{
\item{s}{Spectral data in the format of visa's Spectra object, spectra.data.frame or spectra.matrix.}

\item{b1}{A integer number which defines the wavelength of the 1st spectral band.}

\item{b2}{A integer number which defines the wavelength of the 2nd spectral band.}
}
\value{
\item{ndvi}{The returned values are the new NDVI}
}
\description{
This function calculates a 2-band NDVI using the \code{\link{nsr}} function.
}
\examples{
\dontrun{
library(visa)
s <- NSpec.DF$Spec
ndvi2(x, 780, 680)
}
}
