% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specdf.R
\docType{data}
\name{NSpec.DF}
\alias{NSpec.DF}
\alias{Data-SpectraDataFrame}
\title{Example data in the SpectraDataFrame format}
\format{
A data frame with 19 rows and 2 variables:

\describe{
  \item{N}{Plant nitrogen content}
  \item{spectra}{A variable of Matrix of plant spectra}
  ...
}
}
\usage{
NSpec.DF
}
\description{
A dataset containing the plant Nitrogen content and spectra.
The Spectra matrix is stored as a variable (in a column) of a data.frame.
}
\examples{
library(visa)
data(NSpec.DF)
str(NSpec.DF)
}
\seealso{
\link{data.frame} and \code{\link{NSpec.DB}}
}
\keyword{datasets}
