% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specls.R
\docType{data}
\name{NSpec.LS}
\alias{NSpec.LS}
\alias{Data-Spectra}
\title{Example data in the Spectra class format.}
\format{A Spectra object with 19 rows and 4 slots (spectra, wavelength, s.id, w.unit).

\describe{
  \item{N}{Plant nitrogen content}
  \item{spectra}{A variable of Matrix of plant spectra}
  \item{s.id}{A vector of spectra IDs}
  \item{w.unit}{A character string of wavelength unit ("nm")}
  ...
}}
\source{
\url{visa}
}
\usage{
NSpec.LS
}
\description{
A dataset containing the plant Nitrogen content and spectra. Spectra matrix is stored as a variable of data frame.
}
\examples{
data(NSpec.LS)
str(NSpec.LS)
}
\keyword{datasets}
