% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R, R/plot.lmfit.R, R/plot.spectra.R
\name{plot-method}
\alias{plot-method}
\alias{plot.lmfit}
\alias{plot.spectra}
\title{Plot functions}
\usage{
\method{plot}{lmfit}(x, y)

\method{plot}{spectra}(S, w = wavelength(S), type = "l", xlab = NULL,
  ylab = NULL, lwd = 1:3, col = 1:3, xaxt = "n",
  show.xtick.text = TRUE, xtick = seq(400, 2500, 100),
  rug.interval = 50, ...)
}
\arguments{
\item{x}{A numeric variable}

\item{y}{A numeric variable}

\item{S}{A matrix of spectral data.}

\item{w}{A numeric vector for wavelength.}

\item{type}{Inheritated from plot 'type'}

\item{xlab}{X label.}

\item{ylab}{Y label.}

\item{lwd}{linewidth}

\item{col}{color}

\item{xaxt}{plot function}

\item{show.xtick.text}{whether show x-ticks default true}

\item{xtick}{Define x-ticks labels, default seq(400,2500,100)}

\item{rug.interval}{Define the interval of rugs for wavelength.}

\item{\dots}{Arguments passed to \code{\link{matplot}}}
}
\description{
This functions plots model fit using ggplot.

This function is to make a nice plot of hyperspectral data, e.g., ASD spectrometer,
by using the \code{matplot} function. It customizes x ticks and labels to allow a
better inspection on the spectral signials at each wavelength.
}
\details{
Visualization of linear fit (y = ax + b), using scatter plots and with regression line, as well as
added details of regression equation and R^2.
}
\examples{
\dontrun{
library(visa)
x <- 1:10
y <- 2:11+0.5
plot.lmfit(x,y)
}
\dontrun{
s <- spectra(NSpec.DB)
w <- wavelength(NSpec.DB)
plot.spec(s, w)
}
}
\seealso{
\code{\link{rug}}
}
