% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fit.R
\name{plot.fit}
\alias{plot.fit}
\title{Plot model fit with equation and R2.}
\usage{
plot.fit(x, y)
}
\arguments{
\item{x}{A numeric variable}

\item{y}{A numeric variable}
}
\description{
this functions plots model fit using ggplot.
}
\examples{
\dontrun{
library(visa)
x <- 1:10
y <- 2:11+0.5
plot.fit(x,y)
}
}
