% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spec.R
\name{plot.spec}
\alias{plot.spec}
\title{Plot ASD Hyperspectral Data}
\usage{
\method{plot}{spec}(x, y, type = "l", xlab = NULL, ylab = NULL,
  lwd = 1:3, col = 1:3, xaxt = "n", show.xtick.text = TRUE,
  xtick = seq(400, 2500, 100), rug.interval = 50, ...)
}
\arguments{
\item{x}{A numeric vector for wavelength.}

\item{y}{A matrix of spectral data.}

\item{type}{Inheritated from plot 'type'}

\item{xlab}{X label.}

\item{ylab}{Y label.}

\item{lwd}{linewidth}

\item{col}{color}

\item{xaxt}{plot function}

\item{show.xtick.text}{whether show x-ticks default true}

\item{xtick}{Define x-ticks labels, default seq(400,2500,100)}

\item{rug.interval}{Define the interval of rugs for wavelength.}

\item{\dots}{Arguments passed to \code{\link{matplot}}}
}
\description{
This function is to make a nice plot of hyperspectral data, e.g., ASD spectrometer,
by using the \code{matplot} function. It customizes x ticks and labels to allow a
better inspection on the spectral signials at each wavelength.
}
\seealso{
\code{\link{rug}}
}
