% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specDF.R
\docType{data}
\name{specDF}
\alias{specDF}
\title{Spectral data.}
\format{A data frame with 19 rows and 2 variables:
\describe{
  \item{N}{Plant nitrogen content}
  \item{Spec}{A variable of Matrix of plant spectra}
  ...
}}
\source{
\url{unkow}
}
\usage{
specDF
}
\description{
A dataset containing the plant Nitrogen content and spectra.
}
\keyword{datasets}
