% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\name{wavelength}
\alias{wavelength}
\alias{spectra}
\alias{get.spectra}
\title{Functions to access data of the Spectra object.}
\usage{
wavelength(object, ...)

spectra(object, ...)
}
\arguments{
\item{object}{A object of Spectra}

\item{...}{Other options (... T/F with unit)}

\item{object}{A Spectra object}
}
\description{
Functions to access data of the Spectra object.
}
\examples{
\dontrun{
wavelength(NSpec_List)
}
data(NSpec_List)
spectra_matrix <- spectra(NSpec_List)
}
