% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\name{as.spectra.matrix}
\alias{as.spectra.matrix}
\title{Create a SpectraMatrix Object}
\usage{
as.spectra.matrix(
  spectra = matrix(0),
  wavelength = numeric(0),
  w.unit = character(0)
)
}
\arguments{
\item{spectra}{A matrix.}

\item{wavelength}{A numeric vector.}

\item{w.unit}{A character string.}
}
\value{
A matrix with column names set to the wavelengths.
}
\description{
The constructor \code{as.spectra.matrix} creates a \code{Spectra.Matrix} object.
}
\examples{
smatrix <- as.spectra.matrix(matrix(1:10, 1), 1:10, "nm")
str(smatrix)
}
