% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraDataFrame-class}
\alias{SpectraDataFrame-class}
\title{Class Spectra extended with associated vegetation data}
\description{
Class Spectra extended with associated vegetation data
}
\section{Slots}{

\describe{
\item{\code{entry}}{A data.frame}

\item{\code{spectra}}{A matrix}

\item{\code{wavelength}}{A numeric vector}

\item{\code{s.id}}{A vector}

\item{\code{w.unit}}{A character}
}}

