% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraDatabase}
\alias{SpectraDatabase}
\alias{Spectra}
\title{Class 'SpectraDatabase'}
\description{
SpectraDatabase is an extended 'Spectra' class, with associated vegetation data ('data')
in a \link{data.frame}.
}
\section{Slots}{

\describe{
\item{\code{spectra}}{A matrix}

\item{\code{wavelength}}{A numeric vector}

\item{\code{w.unit}}{A character string}

\item{\code{data}}{A data.frame of vegetation data corresponding to the spectra}
}}

\examples{
new("SpectraDatabase", matrix(1:100,4), 1:25, "nm", data.frame(data=letters[1:4]))
}
