% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\name{as.spectra.dataframe}
\alias{as.spectra.dataframe}
\alias{as.specdf}
\title{Create a SpectraDataframe}
\usage{
as.spectra.dataframe(
  spectra = matrix(0),
  wavelength = numeric(0),
  w.unit = character(0),
  data = data.frame(0),
  ...
)
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{w.unit}{A character string}

\item{data}{A data.frame}

\item{...}{Other options for similar format of variables}
}
\value{
\item{sdf}{Returns a SpectraDataframe.}
}
\description{
Constructor \code{as.spectra.dataframe} function creates a SpectraDataframe object, which is equivalent to the use of \link{as.specdf}.
}
\examples{
sdf <- as.spectra.dataframe(matrix(1:10, 1), 1:10, "nm", data.frame(a = 1, b =2))
str(sdf)
}
