% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraMatrix-class}
\alias{SpectraMatrix-class}
\alias{SpectraMatrix}
\alias{spectra.matrix}
\title{Class 'SpectraMatrix'}
\description{
\code{SpectraMatrix} is an extended \code{Spectra} class.
}
