% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\name{as.spectra.matrix}
\alias{as.spectra.matrix}
\title{Construct a SpectraMatrix object}
\usage{
as.spectra.matrix(spectra = matrix(0), wavelength = numeric(0),
  s.id = numeric(0), w.unit = character(0))
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{s.id}{A numeric vector}

\item{w.unit}{A character}
}
\description{
This function creates a SpectraMatrix object.
}
\examples{
smatr <- as.spectra.matrix(matrix(1:10, 1), 1:10, data.frame(a = 1, b =2), "nm")
str(smatr)
}
