% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visa.R
\docType{package}
\name{visa}
\alias{visa}
\alias{visa-package}
\title{visa: A package for vegetation imaging spectroscopy (e.g. hyperspectral) analysis.}
\description{
The visa package provides a set of functions for hyperspectral analysis, built on other r packages:
}
\details{
It implement two categories of important functions:

\itemize{
\item vegetation indices, and
\item multivariate analyses....
}

To know more about visa, check the vignettes:
`browseVignettes("visa")`
}
\section{Vegetation indices}{

The vegetation indices ...
}

\section{Multivariate analysis}{

The multivariate analysis use the data.frame format...
}

\author{
Kang Yu
}
