% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\docType{methods}
\name{wavelength}
\alias{wavelength}
\alias{waveband}
\alias{wavelength,Spectra-method}
\alias{wavelength,Spectra,ANY-method}
\alias{wavelength,data.frame-method}
\alias{wavelength,data.frame,ANY-method}
\alias{wavelength,matrix-method}
\alias{wavelength,matrix,ANY-method}
\title{Access the wavelength of Spectra}
\usage{
wavelength(object, ...)

\S4method{wavelength}{Spectra}(object, ...)

\S4method{wavelength}{data.frame}(object, ...)

\S4method{wavelength}{matrix}(object, ...)
}
\arguments{
\item{object}{A object of Spectra}

\item{...}{Other options (... T/F with unit)}
}
\description{
Construct generic functions for the Spectra object, spectra.data.frame, and spectra.matrix.
}
\details{
A call to {new} returns a newly allocated object from the class identified by the first argument.
This call in turn calls the method for the generic function `initialize`.
Construct a Spectra class by using the
}
\examples{
library(visa)
# For S4 class Spectra
wavelength(NSpec.DB)
# For spectra data.frame format
wavelength(NSpec.DF)

}
