% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.lmfit.R
\name{ggplot.lmfit}
\alias{ggplot.lmfit}
\title{Plot Model Fit with Equation and R^2.}
\usage{
\method{ggplot}{lmfit}(x, y, environment = parent.frame())
}
\arguments{
\item{x}{A numeric variable}

\item{y}{A numeric variable}
}
\description{
This functions plots model fit using ggplot.
}
\details{
Visualization of linear fit (y = ax + b), using scatter plots and with regression line, as well as
added details of regression equation and R^2.
}
\examples{
\dontrun{
library(visa)
x <- 1:10
y <- 2:11+0.5
ggplot.lmfit(x,y)
}
}
