% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi2.R
\name{ndvi2}
\alias{ndvi2}
\title{Calculate and plot the two NDVI and best NDVI calculated using \code{\link{ndvi}}}
\usage{
ndvi2(x, y, w, p)
}
\arguments{
\item{x}{A matrix of spectral data, a row is a spectrum across all spectral bands.}

\item{y}{A vector.}

\item{w}{A vector of wavelength.}

\item{p}{A vector of two elements (the pair of bands used for NDVI calculation).}
}
\description{
Calculate and plot the two NDVI and best NDVI calculated using \code{\link{ndvi}}
}
\details{
This function plots the best NDVI against the y
}
\examples{
  \\donotrun{
library(visa)
y <- NSpec.DF$N
x <- NSpec.DF$Spec
w <- as.numeric(gsub("nm", "", colnames(NSpec.DF$Spec)))
p <- c(440,445)
ndvi2(x,y,w,p)
wavelength(spectra)
}
}
