% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specdf.R
\docType{data}
\name{NSpec_DF}
\alias{NSpec_DF}
\title{Example data in the Spectral data.frame. format}
\format{A data frame with 19 rows and 2 variables:

\describe{
  \item{N}{Plant nitrogen content}
  \item{spectra}{A variable of Matrix of plant spectra}
  ...
}}
\source{
\url{visa}
}
\usage{
NSpec_DF
}
\description{
A dataset containing the plant Nitrogen content and spectra.
The Spectra matrix is stored as a variable (in a column) of a data.frame.
}
\seealso{
\link{data.frame}
}
\keyword{datasets}
