% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R, R/plot.lmfit.R
\name{plot-method}
\alias{plot-method}
\alias{plot.lmfit}
\title{Plot functions}
\usage{
\method{plot}{lmfit}(x, y)
}
\arguments{
\item{x, y}{numeric vectors.}

\item{y}{A numeric variable}

\item{x}{A numeric variable}
}
\description{
This functions plots model fit using ggplot.
}
\details{
Visualization of linear fit (y = ax + b), using scatter plots and with regression line, as well as
added details of regression equation and R^2.
}
\examples{
\dontrun{
library(visa)
x <- 1:10
y <- 2:11+0.5
plot.lmfit(x,y)
}
}
