% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{Spectra}
\alias{Spectra}
\alias{as.spectra}
\alias{as.spectra.entry}
\title{Construct a Spectra object (S4)}
\usage{
as.spectra(spectra = matrix(0), wavelength = numeric(0),
  s.id = numeric(0), w.unit = character(0), ...)

as.spectra.entry(entry = data.frame(0), spectra = matrix(0),
  wavelength = numeric(0), s.id = numeric(0), w.unit = character(0),
  ...)
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{s.id}{A numeric vector}

\item{w.unit}{A character}

\item{...}{Other parameters}

\item{entry}{A data.frame}

\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{s.id}{A numeric vector}

\item{w.unit}{A character}

\item{entry}{A data.frame}

\item{...}{Other parameters}
}
\description{
This function create a Spectra object
}
\examples{
\dontrun{
s <- as.spectra(matrix(1:100,4), 1:25, 1:4, "nm")
s <- as.spectra.entry(data.frame(x = letters[1:4]), matrix(1:100, 4), 1:25, 1:4, "nm")
}
}
