% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{Spectra}
\alias{Spectra}
\alias{SpectraDatabase}
\alias{as.spectra}
\alias{as.spectra.database}
\title{Create a Spectra or SpectraDatabase}
\usage{
as.spectra(spectra = matrix(0), wavelength = numeric(0), s.id = NULL,
  w.unit = "nm", data = data.frame(), ...)
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{s.id}{A vector}

\item{w.unit}{A character string}

\item{data}{A data.frame}

\item{...}{Other parameters}
}
\description{
This function create a Spectra object.
}
\examples{
\dontrun{
new("Spectra", matrix(1:100,4), 1:25, 1:4, "nm")
s <- as.spectra(matrix(1:100,4), 1:25, 1:4, "nm")
s <- as.spectra.database(matrix(1:100, 4), 1:25, 1:4, "nm", data.frame(x = letters[1:4]))
}
}
