% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R, R/wavelength.R
\name{spectra}
\alias{spectra}
\alias{wavelength}
\title{access the wavelength of Spectra}
\usage{
spectra(object, ...)

wavelength(object, ...)
}
\arguments{
\item{object}{A Spectra object, spectra.data.frame, or spectra.matrix.}

\item{...}{Other options}
}
\description{
Construct generic functions for the Spectra object, spectra.data.frame, and spectra.matrix.
}
\details{
A call to {new} returns a newly allocated object from the class identified by the first argument.
This call in turn calls the method for the generic function `initialize`.
Construct a Spectra class by using the
}
\examples{
\dontrun{
# for the Spectra class
data(NSpec.DB)
spectra_matrix <- spectra(NSpec.DB)
# for spectra data.frame
data(NSpec.DF)
spectra_matrix <- spectra(NSpec.DF)
}
\dontrun{
# for Spectra format
wavelength(NSpec.DB)
# for spectra data.frame format
wavelength(NSpec.DF)
}
}
