% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\name{wavelength}
\alias{wavelength}
\alias{get.wavelength}
\alias{spectra}
\title{Functions to access slot data of the Spectra object.}
\usage{
wavelength(object, ...)

spectra(object, ...)
}
\arguments{
\item{object}{A object of Spectra}

\item{...}{Other options (... T/F with unit)}

\item{object}{A Spectra object}

\item{...}{Other options}
}
\description{
Functions to access slot data of the Spectra object.
}
\examples{
\dontrun{
# for Spectra format
wavelength(NSpec.LS)
# for Spectra data.frame format
wavelength(NSpec.DS)
}
# for the Spectra class
data(NSpec_List)
spectra_matrix <- spectra(NSpec_List)
# for spectra.data.frame
data(NSpec_DF)
spectra_matrix <- spectra(NSpec_DF)

}
