% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R, R/wavelength.R
\name{spectra}
\alias{spectra}
\alias{wavelength}
\title{access the wavelength of Spectra}
\usage{
spectra(object, ...)

wavelength(object, ...)
}
\arguments{
\item{object}{A Spectra object, spectra.data.frame, or spectra.matrix.}

\item{...}{Other options}

\item{object}{A object of Spectra}

\item{...}{Other options (... T/F with unit)}
}
\description{
Construct generic functions for the Spectra object, spectra.data.frame, and spectra.matrix.
}
\details{
A call to {new} returns a newly allocated object from the class identified by the first argument.
This call in turn calls the method for the generic function `initialize`.
Construct a Spectra class by using the
}
\examples{
\dontrun{
# for the Spectra class
data(NSpec_List)
spectra_matrix <- spectra(NSpec_List)
# for spectra.data.frame
data(NSpec_DF)
spectra_matrix <- spectra(NSpec_DF)
}
\dontrun{
# for Spectra format
wavelength(NSpec.LS)
# for Spectra data.frame format
wavelength(NSpec.DS)
}
}
