% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra}
\alias{spectra}
\title{Access the spectra data of Spectra.}
\usage{
spectra(object, ...)
}
\arguments{
\item{object}{A Spectra object, spectra.data.frame, or spectra.matrix.}

\item{...}{Other options}
}
\description{
Functions to access slot data of the Class Spectra.
}
\details{
Construct generic functions for the Spectra object, spectra.data.frame, and spectra.matrix.
A call to {new} returns a newly allocated object from the class identified by the first argument.
This call in turn calls the method for the generic function `initialize`.
}
\examples{
\dontrun{
# For the S4 class 'Spectra'
data(NSpec.DB)
spectra_matrix <- spectra(NSpec.DB)
# For spectra data.frame
data(NSpec.DF)
spectra_matrix <- spectra(NSpec.DF)
}
}
