% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraDatabase-class}
\alias{SpectraDatabase-class}
\alias{SpectraDatabase-class,spectra.database}
\title{Class 'SpectraDatabase'}
\description{
SpectraDatabase is an extended 'Spectra' class, with associated vegetation data ('data')
in a \link{data.frame}.
}
\section{Slots}{

\describe{
\item{\code{spectra}}{A matrix}

\item{\code{wavelength}}{A numeric vector}

\item{\code{w.unit}}{A character string}

\item{\code{data}}{A data.frame of vegetation data corresponding to the spectra}
}}

