% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraDataFrame}
\alias{SpectraDataFrame}
\alias{spectra.data.frame}
\title{Class 'SpectraDataFrame'}
\description{
SpectraDataFrame is an extended 'Spectra' class, with associated vegetation data ('data')
in a \link{data.frame} or \link{list}.
}
\section{Slots}{

\describe{
\item{\code{spectra}}{A matrix}

\item{\code{wavelength}}{A numeric vector}

\item{\code{s.id}}{A vector}

\item{\code{w.unit}}{A character string}

\item{\code{data}}{A data.frame of vegetation data corresponding to the spectra}
}}

\examples{
new("SpectraDataFrame", matrix(1:100,4), 1:25, 1:4, "nm", data.frame(data=letters[1:4]))

}
