% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraMatrix-class}
\alias{SpectraMatrix-class}
\alias{SpectraMatrix-class,}
\alias{spectra.maxtrix}
\alias{as.spectra.matrix}
\title{Class 'SpectraMatrix'}
\usage{
as.spectra.matrix(
  spectra = matrix(0),
  wavelength = numeric(0),
  w.unit = character(0)
)
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{w.unit}{A character string}
}
\value{
\item{sdf}{Returns a SpectraDataframe.}
}
\description{
SpectraMatrix is a extended 'Spectra' class.

Constructor \code{as.spectra.matrix} creates a SpectraMatrix object.
}
\examples{
smatrix <- as.spectra.matrix(matrix(1:10, 1), 1:10, "nm")
str(smatrix)
}
