% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{SpectraMatrix}
\alias{SpectraMatrix}
\alias{as.spectra.matrix}
\title{Class 'SpectraMatrix'}
\usage{
as.spectra.matrix(spectra = matrix(0), wavelength = numeric(0),
  s.id = NULL, w.unit = character(0))
}
\arguments{
\item{spectra}{A matrix}

\item{wavelength}{A numeric vector}

\item{s.id}{A vector of spectral IDs, can be numeric or character}

\item{w.unit}{A character string}
}
\description{
This function creates a SpectraMatrix object.
}
\examples{
smatrix <- as.spectra.matrix(matrix(1:10, 1), 1:10, 1, "nm")
str(smatrix)
}
