% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{sr}
\alias{sr}
\alias{nsr}
\title{Calculate simple ratio (SR) indices.}
\usage{
sr(x, y)

nsr(x, y)
}
\arguments{
\item{x, y}{numeric vectors.}
}
\description{
descriptions

descriptions
}
\details{
details
}
