% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visa.R
\docType{package}
\name{visa}
\alias{visa}
\alias{visa-package}
\title{visa: Vegetation Imaging Spectroscopy Analyzer}
\description{
For an easy-to-use tool for working with Hyperspectral Remote Sensing (HRS, also known as Imaging Spectroscopy) data.
}
\details{
The visa package provides a set of functions for hyperspectral data analysis, with the main idea
of simplifying the analysis and interpretation of vegetation spectral data.

It implement two categories of important functions, including its own function and also
some built on other r packages.

\itemize{
\item vegetation indices, and
\item multivariate analyses....
}

To know more about visa, check the vignettes:
`browseVignettes("visa")`
}
\section{Vegetation Indices}{

Support the calculation of vegetation indices from literature, as well as creation of new indices.
}

\section{Multivariate analysis}{

... not implemented yet
The multivariate analysis use the spectra.data.frame format.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kang-yu/visa}
  \item Report bugs at \url{https://github.com/kang-yu/visa/issues}
}

}
\author{
Kang Yu
}
\keyword{internal}
