% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\name{wavelength}
\alias{wavelength}
\title{Access the wavelength of Spectra}
\usage{
wavelength(object, ...)
}
\arguments{
\item{object}{A Spectra object, spectra.data.frame, or spectra.matrix.}

\item{...}{Other options.}
}
\description{
Construct generic functions for the Spectra object, spectra.data.frame, and spectra.matrix.
}
\details{
A call to {new} returns a newly allocated object from the class identified by the first argument.
This call in turn calls the method for the generic function `initialize`.
Construct a Spectra class by using the
}
\examples{
\dontrun{
# For S4 class Spectra
wavelength(NSpec.DB)
# For spectra data.frame format
wavelength(NSpec.DF)
}
}
